java

+++ boot_camp/src/javabasico/operadores/aula01/MinhaClass.java
@@ -7,7 +7,7 @@ public class MinhaClass {
         int multiplicacao = 20 * 7;
         int divisao = 15 / 3;
         int modulo = 18 % 3;
-        double resultado = (10 * 7) + (20/4);
+        double resultado = (10 * 7) + ((double) 20 /4);
 
         System.out.println(soma);
         System.out.println(subtracao);
